refactor: renamed security classes to ApiKey*
